Hoist staged transaction utils
