+++ packages/user/FractalCore/ui/src/lib/graphql/fractals/getFractal.ts
@@ -1,10 +1,10 @@
 import { z } from "zod";
 
 import { FRACTALS_SERVICE } from "@/lib/constants";
-import { zDateTime } from "@/lib/zod/DateTime";
 
 import { graphql } from "@shared/lib/graphql";
 import { Account, zAccount } from "@shared/lib/schemas/account";
+import { zDateTime } from "@shared/lib/schemas/date-time";
 import { zU8 } from "@shared/lib/schemas/u8";
 
 export const zFractal = z
