+++ packages/user/FractalCore/ui/src/lib/graphql/fractals/getGuildApplications.ts
@@ -1,10 +1,10 @@
 import { z } from "zod";
 
 import { FRACTALS_SERVICE } from "@/lib/constants";
-import { zDateTime } from "@/lib/zod/DateTime";
 
 import { graphql } from "@shared/lib/graphql";
 import { Account, zAccount } from "@shared/lib/schemas/account";
+import { zDateTime } from "@shared/lib/schemas/date-time";
 
 export const zGuildApplicationListInstance = z.object({
     member: zAccount,
